telegram-bot/requirements.txt
python-telegram-bot
selenium
python-dotenv
